Fixed the BUG(S)!
it was not an angle bug, neither it was gaussians' software problem. (1) Major bug: It was a minus sign, coming from defining vectors rij = rj-ri instead of rij = rj-ri. Intuitively the force goes from i to j if the gradient is positive. However, that is why there is a minus sign in the gradient: fi = -dudr*dudri. Otherwise fi = dudr*(rj-ri). Draw on a paper a schetch and see. So the bug was coming from the fact that by defining rij=rj-ri the minus sign was counted twice! (2) Major bug: Probably why I thought it was an angle bug. The gaussian software makes some reflection transformations on the data. So instead of reading the standard orientation we should read the 'Input orientation'. Fixed that . (3) Logic bug, Forces are orders of magnitude (2-3) higher than energies. It is better to normalize data during the training process. Now both the training and the development set are normalized with the default option ' normalize_data = 1 ' Will check also on Ag clusters to make sure everything is ok!